пытаюсь удалить обработчики при закрытии окна с загрузкой изображения
