+++ index.html
@@ -10,6 +10,7 @@
   <script src="./vendor/pristine/pristine.min.js"></script>
 <script src="./js/form/form-modal.js" type="module"></script>
 <script src="./js/form/form-validate.js" type="module"></script>
+<script src="./js/form/form-scale.js" type="module"></script>
 
   <title>Кекстаграм</title>
 </head>
